refactor: move image server utilities
