+++ lib/server/images/get-image-dimensions.ts
